+++ nata-monkey/src/main/java/com/nata/action/Action.java
@@ -1,61 +1,33 @@
 package com.nata.action;
 
-import java.util.Objects;
-
 /**
  * Author: Calvin Meng
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-13 16:36
  */
 public abstract class Action {
-    private String name = null;
+    protected String type;
     protected int count = 0;
-    private final double RINIT= 2.0;
+    protected final double BASE = 2.0;
 
-
-    public Action() {
-        this.name = ActionType.UNKNOWN;
+    public Action(String type) {
+        this.type = type;
     }
 
-    public double getReward(){
-        if(count == 0){
-            return RINIT;
-        }else{
-            return 1.0/count;
-        }
-    }
-
-    protected void setName(String name){
-       this.name = name;
-    }
-
+    /**
+     * Implement how to fire the action
+     */
     abstract public void fire();
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
+    /**
+     * Give the rule how to reward the action
+     * @return reward value
+     */
+    abstract public double getReward();
 
-    @Override
-    public boolean equals(Object otherObject) {
-        if (this == otherObject) {
-            return true;
-        }
-
-        if (otherObject == null) {
-            return false;
-        }
-
-        if (getClass() != otherObject.getClass()) {
-            return false;
-        }
-
-        Action other = (Action) otherObject;
-        return Objects.equals(name, other.name) ;
-    }
 
     @Override
     public String toString() {
-        return "[Action] ";
+        return "[Action]("+type+","+count +")";
     }
 }