+++ nata-monkey/src/main/java/com/nata/action/ActionType.java
@@ -6,15 +6,15 @@ package com.nata.action;
  * Update: 2016-01-13 20:58
  */
 public class ActionType {
-    public static final String UNKNOWN= "unknown";
-    public static final String TAP = "tap";
-    public static final String LONGCLICK= "longclick";
+    public static final String UNKNOWN= "UnknownAction";
+    public static final String TAP = "TapAction";
+    public static final String LONG_CLICK= "LongClickAction";
 
-    public static final String SWIPE= "swipe";
-    public static final String INPUT= "input";
-    public static final String BACK= "back";
-    public static final String MENU= "menu";
-    public static final String STARTAPP= "start app";
+    public static final String SWIPE= "SwipeAction";
+    public static final String INPUT= "TextInputAction";
+    public static final String BACK= "BackAction";
+    public static final String MENU= "MenuAction";
+    public static final String START_APP= "StartAppAction";
 
 
 }