+++ nata-monkey/src/main/java/com/nata/action/LongClickAction.java
@@ -1,7 +1,7 @@
 package com.nata.action;
 
 import com.nata.cmd.AdbDevice;
-import com.nata.element.Element;
+import com.nata.element.Widget;
 
 /**
  * Author: Calvin Meng
@@ -9,33 +9,42 @@ import com.nata.element.Element;
  * Update: 2016-03-24 21:23
  */
 public class LongClickAction extends Action{
-
     private AdbDevice device = null;
-    private Element element = null;
-    public LongClickAction(AdbDevice device,Element element){
-        super.setName(ActionType.LONGCLICK);
+    private Widget widget = null;
+
+    private final double LONG_CLICK_REWARD = BASE + 0.2;
+
+    public LongClickAction(AdbDevice device,Widget widget){
+        super(ActionType.LONG_CLICK);
         this.device = device;
-        this.element = element;
+        this.widget = widget;
     }
 
     @Override
     public void fire() {
-        device.longPress(element.getX(),element.getY());
+        device.longPress(widget.getX(), widget.getY());
         count++;
+    }
+
+    @Override
+    public double getReward() {
+        if(count == 0){
+            return LONG_CLICK_REWARD;
+        }else{
+            return 1.0/count;
+        }
     }
 
 
     @Override
     public String toString() {
-        return super.toString()+"LongClickAction{" +
-                "element=" + element +
-                '}';
+        return super.toString()+"@"+widget;
     }
 
     @Override
     public int hashCode() {
         int hash = 17;
-        hash = 31 * hash + element.hashCode();
+        hash = 31 * hash + widget.hashCode();
         return hash;
     }
 
@@ -55,7 +64,7 @@ public class LongClickAction extends Action{
 
         LongClickAction other = (LongClickAction) otherObject;
 
-        if(!element.equals(other.element)){
+        if(!widget.equals(other.widget)){
             return false;
         }