+++ nata-monkey/src/main/java/com/nata/element/Widget.java
@@ -9,7 +9,7 @@ import java.util.regex.Pattern;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-13 16:46
  */
-public class Element {
+public class Widget {
 //    private Activity activity;
     private int X;
     private int Y;
@@ -17,7 +17,7 @@ public class Element {
     private String resourceId;
     private String text;
 
-    public Element(UINode node){
+    public Widget(UINode node){
         this.className = node.getClassName();
         this.resourceId = node.getResourceId();
         this.text = node.getText();
@@ -44,7 +44,7 @@ public class Element {
     @Override
     public String toString() {
         String description = className + " " + resourceId +" " + text;
-        return "Element{" +
+        return "Widget{" +
                 "X=" + X +
                 ", Y=" + Y +
                 ", description='" + description + '\'' +
@@ -54,11 +54,8 @@ public class Element {
     @Override
     public int hashCode() {
         int hash = 17;
-        hash = 31 * hash + X;
-        hash = 31 * hash + Y;
         hash = 31 * hash + className.hashCode();
         hash = 31 * hash + resourceId.hashCode();
-        hash = 31 * hash + text.hashCode();
 
         return hash;
     }
@@ -77,23 +74,14 @@ public class Element {
             return false;
         }
 
-        Element other = (Element) otherObject;
-        if(X!= other.X){
-            return false;
-        }
-        if(Y!= other.Y){
-            return false;
-        }
+        Widget other = (Widget) otherObject;
+
         if(!resourceId.equals(other.resourceId)){
             return false;
         }
         if(!className.equals(other.className)){
             return false;
         }
-        if(!text.equals(other.text)){
-            return false;
-        }
-
         return true;
     }