+++ nata-monkey/src/main/java/com/nata/rules/SetUpRule.java
@@ -1,0 +1,9 @@
+package com.nata.rules;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-03 15:22
+ */
+public class SetUpRule {
+}