+++ nata-monkey/src/main/java/com/nata/state/SubState.java
@@ -1,0 +1,9 @@
+package com.nata.state;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-03 13:16
+ */
+public class SubState {
+}